requirements.txt
black==21.11b1
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
python-dotenv==0.19.2
regex==2021.11.10
SQLAlchemy==1.4.27
tomli==1.2.2
typing_extensions==4.0.0
Werkzeug==2.0.2
wincertstore==0.2
